services/intent-service/requirements.txt
# Intent Recognition Service Requirements
# CPU-optimized deployment for Turkish BERT

# Core gRPC dependencies
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Machine Learning dependencies for Turkish BERT
transformers>=4.30.0
torch>=2.0.0
onnxruntime>=1.15.0  # CPU-only inference
numpy>=1.24.0

# Text processing
tokenizers>=0.13.0

# Async support
asyncio

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Logging
structlog>=23.0.0

# Development hot-reload support
watchdog>=3.0.0